+++ sat/cli/sensors/bmc.py
@@ -32,6 +32,7 @@ import json
 import re
 
 
+
 LOGGER = logging.getLogger(__name__)